Secured endpoints, created login test

+++ Backend/src/main/java/org/example/backend/controllers/ArticleController.java
@@ -1,5 +1,6 @@
 package org.example.backend.controllers;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.models.article.Article;
 import org.example.backend.models.article.ArticleCreateDto;
@@ -8,9 +9,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
-// todo: secure endpoints
 @RestController
 @RequestMapping("/api/articles")
 @RequiredArgsConstructor
@@ -30,28 +33,47 @@ public class ArticleController {
 
     // not found exception
     @GetMapping("/{id}")
-    public Article getById(@PathVariable Long id) {
-        return articleService.getById(id);
+    public ResponseEntity<Article> getById(@PathVariable Long id) {
+        try {
+            return ResponseEntity.ok(articleService.getById(id));
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 
-    // todo: secure
     @PostMapping
     public ResponseEntity<Article> create(@RequestBody ArticleCreateDto article) {
-        Article created = articleService.create(article);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth == null || !auth.isAuthenticated()) {
+            throw new AccessDeniedException("Unauthenticated");
+        }
+        Article created = articleService.create(article, (Long) auth.getDetails());
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
 
-    // todo: secure
-    // not found exception
     @PutMapping("/{id}")
-    public Article update(@PathVariable Long id, @RequestBody Article article) {
-        return articleService.update(id, article);
+    public ResponseEntity<Article> update(@PathVariable Long id, @RequestBody Article article) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth == null || !auth.isAuthenticated()) {
+            throw new AccessDeniedException("Unauthenticated");
+        }
+        var userId = (Long)auth.getDetails();
+        if (!articleService.getById(id).getCreatorId().equals(userId)){
+            throw new AccessDeniedException("Unauthenticated");
+        }
+        return ResponseEntity.ok(articleService.update(id, article));
     }
 
-    // todo: secure
-    // not found exception
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth == null || !auth.isAuthenticated()) {
+            throw new AccessDeniedException("Unauthenticated");
+        }
+        var userId = (Long)auth.getDetails();
+        if (!articleService.getById(id).getCreatorId().equals(userId)){
+            throw new AccessDeniedException("Unauthenticated");
+        }
         articleService.delete(id);
         return ResponseEntity.noContent().build();
     }
